Added a whole bunch of algorithms. Renamed some categories. Changed version to 0.11.

+++ NativeMethods.cs
@@ -11,7 +11,7 @@
         private const UInt32 EM_GETCUEBANNER = 0x1502;
 
         [DllImport("user32.dll", CharSet = CharSet.Auto)]
-        private static extern Int32 SendMessage(IntPtr hWnd, UInt32 msg, IntPtr wParam, [MarshalAs(UnmanagedType.LPWStr)] string lParam);
+        private static extern IntPtr SendMessage(IntPtr hWnd, UInt32 msg, IntPtr wParam, [MarshalAs(UnmanagedType.LPWStr)] string lParam);
 
         public static void SetCueText(Control control, string text)
         {
